requirements.txt
elevenlabs>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
soundfile>=0.12.1
numpy>=1.26.0
